% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ET_params.R
\name{add_ET_params}
\alias{add_ET_params}
\title{Wrapper for all ET hourly functions to add to the dataset}
\usage{
add_ET_params(data)
}
\arguments{
\item{data}{dataframe with all the necessary measured values}
}
\value{
dataframe with all the intermediate ET params with the reference ETo and ETr per hour
}
\description{
Wrapper for all ET hourly functions to add to the dataset
}
