% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{beta}
\alias{beta}
\title{Calculate the angle of the sun above the horizon}
\usage{
beta(lat, long, sct, date)
}
\arguments{
\item{lat}{the latitude in degrees, positive for North, negative for East}

\item{long}{the longitude in degrees, positive for West, negative for East}

\item{sct}{the standard clock time}

\item{date}{date}
}
\description{
Calculate the angle of the sun above the horizon
}
