% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2C.R
\name{convert2C}
\alias{convert2C}
\title{Conver to Celsius from Farenheit}
\usage{
convert2C(TF)
}
\arguments{
\item{TF}{numerical value for a temperture in F}
}
\description{
Conver to Celsius from Farenheit
}
