% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2MJm2h.R
\name{convert2MJm2h}
\alias{convert2MJm2h}
\title{Convert from W/m2*h to MJ/m2*h}
\usage{
convert2MJm2h(solar)
}
\arguments{
\item{solar}{measured solar radiation in Watts/meter^2*hour}
}
\value{
solar radiation in MJ/m2*h
}
\description{
Convert from W/m2*h to MJ/m2*h
}
