% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{delta}
\alias{delta}
\title{Calculate the slope of the saturation vapor pressure-temperature curve}
\usage{
delta(Tm)
}
\arguments{
\item{Tm}{mean temperature in C}
}
\description{
Calculate the slope of the saturation vapor pressure-temperature curve
}
