% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_hour_calc.R
\name{ET_hr_calc}
\alias{ET_hr_calc}
\title{Function to calculate hourly evapotranspiration}
\usage{
ET_hr_calc(Tm, z, u2, Tdew, Rn, G_ETo, G_ETr)
}
\arguments{
\item{Tm}{numeric value of mean hourly temperature in F}

\item{z}{numeric value for elevation in meters}

\item{u2}{numeric value for wind speed in m/s}

\item{Tdew}{numeric value for dewpoint temperature in F}

\item{Rn}{numeric value for net radiation in MJ/(m2*h)}

\item{G_ETo}{numeric value for short crop soil heat flux}

\item{G_ETr}{numeric value for tall crop soil heat flux}
}
\value{
list of the short and tall crop reference evapotranspiration
}
\description{
Function to calculate hourly evapotranspiration
}
\keyword{ET,}
\keyword{evapotranspiration}
\keyword{evapotranspiration,}
\keyword{reference}
