% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcd.R
\name{fcd}
\alias{fcd}
\title{Calculate the cloudiness function}
\usage{
fcd(Rs_Rso)
}
\arguments{
\item{Rs_Rso}{the ratio of the measured radiation (Rs) to the clear sky radiation (Rso)}
}
\description{
Calculate the cloudiness function
}
