% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcd_beta_corr.R
\name{fcd_beta_corr}
\alias{fcd_beta_corr}
\title{Correction needed for the fcd for nightime}
\usage{
fcd_beta_corr(fcd_vector)
}
\arguments{
\item{fcd_vector}{a numerical vector of the fcd values}
}
\value{
a numerical vector of corrected fcd
}
\description{
Correction needed for the fcd for nightime
}
