% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G.R
\name{G}
\alias{G}
\title{Calculate the soil heat flux}
\usage{
G(Rn)
}
\arguments{
\item{Rn}{numberical value for the net radiation}
}
\value{
list of soil heat flux for both short and tall reference crops
}
\description{
Calculate the soil heat flux
}
