% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyschro.R
\name{pyschro}
\alias{pyschro}
\title{Calculate the psychrometric constant}
\usage{
pyschro(P, cp = 1013, epsilon = 0.622, lambda = 2.45)
}
\arguments{
\item{P}{numeric value for the atmospheric pressure in kPa}
}
\description{
Calculate the psychrometric constant
}
