% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ra_hr.R
\name{Ra_hr}
\alias{Ra_hr}
\title{Calculate the extraterrestrail radiation}
\usage{
Ra_hr(lat, date, dr, w1, w2, Gsc = 4.92)
}
\arguments{
\item{lat}{latitude in degrees}

\item{dr}{distance from earth to the sun}

\item{w1}{solar time angle for beginning of hour}

\item{w2}{solar time angle for end of hour}
}
\description{
Calculate the extraterrestrail radiation
}
