% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnl.R
\name{Rnl}
\alias{Rnl}
\title{Calculate long wave radiation}
\usage{
Rnl(fcd, Tdew, Tm)
}
\arguments{
\item{fcd}{numerical value for cloudiness function}

\item{Tdew}{mean hourly dewpoint temperature in C}

\item{Tm}{mean hourly temperature in C}
}
\description{
Calculate long wave radiation
}
