% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rs_Rso_correction.R
\name{Rs_Rso_correction}
\alias{Rs_Rso_correction}
\title{Correction for the Rs/Rso ratio}
\usage{
Rs_Rso_correction(Rs_Rso)
}
\arguments{
\item{Rs_Rso}{raio of the measured and clear sky radiation}
}
\description{
Correction for the Rs/Rso ratio
}
