% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rso.R
\name{Rso}
\alias{Rso}
\title{Calculate clear sky radiation}
\usage{
Rso(z, lat, date, dr, w1, w2)
}
\arguments{
\item{z}{elevation in meters}

\item{lat}{latitude in degrees}

\item{dr}{distance from earth to the sun}

\item{w1}{solar time angle for beginning of hour}

\item{w2}{solar time angle for end of hour}
}
\description{
Calculate clear sky radiation
}
