% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_time_angle.R
\name{solar_time_angle}
\alias{solar_time_angle}
\title{Calculate the solar time angle}
\usage{
solar_time_angle(date, long, sct, Lz = 120)
}
\arguments{
\item{long}{the longitude in degrees, positive for West, negative for East}

\item{sct}{the standard clock time}
}
\description{
Calculate the solar time angle
}
