% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunset_angle.R
\name{sunset_angle}
\alias{sunset_angle}
\title{Calculate the sunset hour angle of the sun}
\usage{
sunset_angle(lat, date)
}
\arguments{
\item{lat}{latitude in degrees, postive for N, negative for S}
}
\description{
Calculate the sunset hour angle of the sun
}
