% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w1.R
\name{w1}
\alias{w1}
\title{Solar time angles at the beginning and end of each hourly period}
\usage{
w1(w)
}
\arguments{
\item{w}{solar time angle}
}
\description{
Solar time angles at the beginning and end of each hourly period
}
