% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w1_correction.R
\name{w1_correction}
\alias{w1_correction}
\title{Hourly solar time angle correction}
\usage{
w1_correction(w1, w2, Ws, neg_Ws)
}
\arguments{
\item{w1}{solar time angle at the beginning of the hour}

\item{w2}{solar time angle at the end of the hour}

\item{Ws}{solar time angle}

\item{neg_Ws}{sunrise solar time angle}
}
\description{
Hourly solar time angle correction
}
