% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2.R
\name{w2}
\alias{w2}
\title{Solar time angles at the beginning and end of each hourly period}
\usage{
w2(w)
}
\arguments{
\item{w}{solar time angle}
}
\description{
Solar time angles at the beginning and end of each hourly period
}
