% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_correction.R
\name{w2_correction}
\alias{w2_correction}
\title{Hourly solar time angle correction}
\usage{
w2_correction(w2, Ws, neg_Ws)
}
\arguments{
\item{w2}{solar time angle at the end of the hour}

\item{Ws}{solar time angle}

\item{neg_Ws}{sunrise solar time angle}
}
\description{
Hourly solar time angle correction
}
